Add files via upload
I can't seem to post a video of the light-dependent resistor because the file is too big.  I've tried to resize the video by cutting some of it out using adobe Premiere Pro but when I do, I format it to AVI or MOV the file becomes bigger.  I've also tried to compress the file but when I do that, it doesn't seem to reduce the file size either.  The best I can do is use a program that converts it into a smaller file but only 1/3 of the full video is converted shown here.  What I can do is show you the video in person in the next class or please let me know how you want me to send it electronically